FM-999 Remove generated API Code
