--- src/main/kotlin/uk/gov/justice/digital/hmpps/approvedpremisesapi/api/cas1/ChangeRequestsCas1Delegate.kt
@@ -1,82 +1,0 @@
-package uk.gov.justice.digital.hmpps.approvedpremisesapi.api.cas1
-
-import org.springframework.http.HttpStatus
-import org.springframework.http.MediaType
-import org.springframework.http.ResponseEntity
-import org.springframework.web.context.request.NativeWebRequest
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.cas1.dto.Cas1ChangeRequest
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.cas1.dto.Cas1NewChangeRequest
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.cas1.dto.Cas1RejectChangeRequest
-import java.util.Optional
-
-@jakarta.annotation.Generated(value = ["org.openapitools.codegen.languages.KotlinSpringServerCodegen"], comments = "Generator version: 7.13.0")
-interface ChangeRequestsCas1Delegate {
-
-  fun getRequest(): Optional<NativeWebRequest> = Optional.empty()
-
-  fun createPlacementAppeal(
-    placementRequestId: java.util.UUID,
-    cas1NewChangeRequest: Cas1NewChangeRequest,
-  ): ResponseEntity<Unit> {
-    getRequest().ifPresent { request ->
-      for (mediaType in MediaType.parseMediaTypes(request.getHeader("Accept"))) {
-        if (mediaType.isCompatibleWith(MediaType.valueOf("application/problem+json"))) {
-          ApiUtil.setExampleResponse(request, "application/problem+json", "Custom MIME type example not yet supported: application/problem+json")
-          break
-        }
-      }
-    }
-    return ResponseEntity(HttpStatus.NOT_IMPLEMENTED)
-  }
-
-  fun createPlacementExtension(
-    placementRequestId: java.util.UUID,
-    cas1NewChangeRequest: Cas1NewChangeRequest,
-  ): ResponseEntity<Unit> {
-    getRequest().ifPresent { request ->
-      for (mediaType in MediaType.parseMediaTypes(request.getHeader("Accept"))) {
-        if (mediaType.isCompatibleWith(MediaType.valueOf("application/problem+json"))) {
-          ApiUtil.setExampleResponse(request, "application/problem+json", "Custom MIME type example not yet supported: application/problem+json")
-          break
-        }
-      }
-    }
-    return ResponseEntity(HttpStatus.NOT_IMPLEMENTED)
-  }
-
-  fun createPlannedTransfer(
-    placementRequestId: java.util.UUID,
-    cas1NewChangeRequest: Cas1NewChangeRequest,
-  ): ResponseEntity<Unit> {
-    getRequest().ifPresent { request ->
-      for (mediaType in MediaType.parseMediaTypes(request.getHeader("Accept"))) {
-        if (mediaType.isCompatibleWith(MediaType.valueOf("application/problem+json"))) {
-          ApiUtil.setExampleResponse(request, "application/problem+json", "Custom MIME type example not yet supported: application/problem+json")
-          break
-        }
-      }
-    }
-    return ResponseEntity(HttpStatus.NOT_IMPLEMENTED)
-  }
-
-  fun get(
-    placementRequestId: java.util.UUID,
-    changeRequestId: java.util.UUID,
-  ): ResponseEntity<Cas1ChangeRequest> {
-    getRequest().ifPresent { request ->
-      for (mediaType in MediaType.parseMediaTypes(request.getHeader("Accept"))) {
-        if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
-          ApiUtil.setExampleResponse(request, "application/json", "{  \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\",  \"requestReason\" : {    \"name\" : \"name\",    \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\"  },  \"decision\" : \"approved\",  \"spaceBookingId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",  \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",  \"type\" : \"placementAppeal\",  \"decisionJson\" : \"{}\",  \"rejectionReason\" : {    \"name\" : \"name\",    \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\"  },  \"requestJson\" : \"{}\",  \"updatedAt\" : \"2000-01-23T04:56:07.000+00:00\"}")
-          break
-        }
-      }
-    }
-    return ResponseEntity(HttpStatus.NOT_IMPLEMENTED)
-  }
-
-  fun reject(
-    placementRequestId: java.util.UUID,
-    changeRequestId: java.util.UUID,
-    cas1RejectChangeRequest: Cas1RejectChangeRequest,
-  ): ResponseEntity<Unit> = ResponseEntity(HttpStatus.NOT_IMPLEMENTED)
-}
